Make the extension non-relocatable, use geoip schema by default
The extension is no longer relocatable. The functions were not using
schema, so it was not working reliably anyway (it would require setting
search_path etc.). So just make it non-relocatable and use @extschema@
in the SQL script.

Also, set the default schema to 'geoip' instead of just using 'public'.
It's still possible to override this in CREATE EXTENSION, of course.