Initial support for current GeoLite2 database
Structure of the GeoLite database changed quite a bit since 2012, and
data files for the old structure are no longer available (or woefully
out of date). So rework the structure to match current files from

  https://dev.maxmind.com/geoip/geoip2/geolite2/

This does support both ipv4 and ipv6 addresses, which is nice, but it
will require more work to get the same performance as older versions.
For example the addresses are no longer represented by integers, so the
comparisons are a bit more expensive.