Optimize the query functions by setting LIMIT 1
We know each IP address can only match a single IP block, but the row
estimates for <<= operator are tricky and may end up being high enough
to plan parallel query. Which is a bad idea, because it adds a lot of
latency without having a chance to improve performance.